ml/training/requirements.txt
torchvision==0.15.2
Pillow==10.0.0
tqdm==4.66.1